Move LayerVariable related functions out of tfjs_backend to the new variable.ts file
