+++ src/backend/tfjs_backend_test.ts
@@ -828,213 +828,6 @@ describeMathCPUAndGPU('tile', () => {
   });
 });
 
-describeMathCPUAndGPU('Create Variable', () => {
-  it('From Tensor, no explicit name', () => {
-    const v = K.variable(zeros([2, 2]));
-    expect(v.name.indexOf('Variable')).toEqual(0);
-    expect(v.shape).toEqual([2, 2]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-
-  it('From Tensor, no explicit name', () => {
-    const v = K.variable(zeros([3]));
-    expect(v.name.indexOf('Variable')).toEqual(0);
-    expect(v.shape).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([0, 0, 0]));
-  });
-
-  it('From Tensor, explicit name', () => {
-    const v = K.variable(zeros([3]), undefined, 'Var1');
-    expect(v.name.indexOf('Var1')).toEqual(0);
-    expect(v.shape).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([0, 0, 0]));
-  });
-});
-
-describeMathCPUAndGPU('batchGetValue', () => {
-  it('Legnth-3 Array, Mixed Tensor and Variable', () => {
-    const v1 = K.variable(zeros([]));
-    const v2 = K.variable(zeros([2]));
-    const v3 = K.variable(zeros([2, 2]));
-    const values = K.batchGetValue([v1, v2, v3]);
-    expect(values.length).toEqual(3);
-    expect(values[0].shape).toEqual([]);
-    expect(values[0].dataSync()).toEqual(new Float32Array([0]));
-    expect(values[1].shape).toEqual([2]);
-    expect(values[1].dataSync()).toEqual(new Float32Array([0, 0]));
-    expect(values[2].shape).toEqual([2, 2]);
-    expect(values[2].dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-});
-
-describeMathCPUAndGPU('batchSetValue', () => {
-  it('Update using Tensor values', () => {
-    const v1 = K.randomUniformVariable([2], 0, 1);
-    const v2 = K.randomUniformVariable([2, 2], 0, 1);
-    K.batchSetValue([[v1, zeros([2])], [v2, zeros([2, 2])]]);
-    expect(v1.shape).toEqual([2]);
-    expect(v1.read().dataSync()).toEqual(new Float32Array([0, 0]));
-    expect(v2.shape).toEqual([2, 2]);
-    expect(v2.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-
-  it('Update using Tensor values', () => {
-    const v1 = K.randomUniformVariable([], 0, 1);
-    const v2 = K.randomUniformVariable([2, 2, 1], 0, 1);
-    K.batchSetValue([[v1, zeros([])], [v2, zeros([2, 2, 1])]]);
-    expect(v1.shape).toEqual([]);
-    expect(v1.read().dataSync()).toEqual(new Float32Array([0]));
-    expect(v2.shape).toEqual([2, 2, 1]);
-    expect(v2.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-
-  it('Update empty Array', () => {
-    K.batchSetValue([]);
-  });
-});
-
-describeMathCPUAndGPU('ZerosVariable', () => {
-  it('Scalar', () => {
-    const s = K.zerosVariable([], DType.float32, 'Scalar');
-    expect(s.name.indexOf('Scalar')).toEqual(0);
-    expect(K.shape(s.read())).toEqual([]);
-    expect(s.read().dataSync()).toEqual(new Float32Array([0]));
-  });
-
-  it('Vector', () => {
-    const v = K.zerosVariable([3], DType.float32, 'Vector');
-    expect(v.name.indexOf('Vector')).toEqual(0);
-    expect(K.shape(v.read())).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([0, 0, 0]));
-  });
-
-  it('Matrix', () => {
-    const m = K.zerosVariable([2, 2], DType.float32, 'Matrix');
-    expect(m.name.indexOf('Matrix')).toEqual(0);
-    expect(K.shape(m.read())).toEqual([2, 2]);
-    expect(m.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-
-  it('3D', () => {
-    const t = K.zerosVariable([2, 2, 2], DType.float32, 'Tertiary');
-    expect(t.name.indexOf('Tertiary')).toEqual(0);
-    expect(K.shape(t.read())).toEqual([2, 2, 2]);
-    expect(t.read().dataSync()).toEqual(new Float32Array([
-      0, 0, 0, 0, 0, 0, 0, 0
-    ]));
-  });
-
-  it('4D', () => {
-    const q = K.zerosVariable([1, 2, 1, 3], DType.float32, 'Quaternary');
-    expect(q.name.indexOf('Quaternary')).toEqual(0);
-    expect(K.shape(q.read())).toEqual([1, 2, 1, 3]);
-    expect(q.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0, 0, 0]));
-  });
-});
-
-describeMathCPUAndGPU('OnesVariable', () => {
-  it('Scalar', () => {
-    const s = K.onesVariable([], DType.float32, 'Scalar');
-    expect(s.name.indexOf('Scalar')).toEqual(0);
-    expect(K.shape(s.read())).toEqual([]);
-    expect(s.read().dataSync()).toEqual(new Float32Array([1]));
-  });
-  it('Vector', () => {
-    const v = K.onesVariable([3], DType.float32, 'Vector');
-    expect(v.name.indexOf('Vector')).toEqual(0);
-    expect(K.shape(v.read())).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([1, 1, 1]));
-  });
-  it('Matrix', () => {
-    const m = K.onesVariable([2, 2], DType.float32, 'Matrix');
-    expect(m.name.indexOf('Matrix')).toEqual(0);
-    expect(K.shape(m.read())).toEqual([2, 2]);
-    expect(m.read().dataSync()).toEqual(new Float32Array([1, 1, 1, 1]));
-  });
-  it('3D', () => {
-    const t = K.onesVariable([2, 2, 2], DType.float32, 'Tertiary');
-    expect(t.name.indexOf('Tertiary')).toEqual(0);
-    expect(K.shape(t.read())).toEqual([2, 2, 2]);
-    expect(t.read().dataSync()).toEqual(new Float32Array([
-      1, 1, 1, 1, 1, 1, 1, 1
-    ]));
-  });
-  it('4D', () => {
-    const q = K.onesVariable([1, 2, 1, 3], DType.float32, 'Quaternary');
-    expect(q.name.indexOf('Quaternary')).toEqual(0);
-    expect(K.shape(q.read())).toEqual([1, 2, 1, 3]);
-    expect(q.read().dataSync()).toEqual(new Float32Array([1, 1, 1, 1, 1, 1]));
-  });
-});
-
-describeMathCPUAndGPU('ZerosLike', () => {
-  it('Scalar', () => {
-    const s = K.zerosLike(tfc.randomUniform([], -10, 10));
-    expect(K.shape(s.read())).toEqual([]);
-    expect(s.read().dataSync()).toEqual(new Float32Array([0]));
-  });
-
-  it('Vector', () => {
-    const v = K.zerosLike(tfc.randomUniform([3], -10, 10));
-    expect(K.shape(v.read())).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([0, 0, 0]));
-  });
-
-  it('Matrix', () => {
-    const m = K.zerosLike(tfc.randomUniform([2, 2], -10, 10));
-    expect(K.shape(m.read())).toEqual([2, 2]);
-    expect(m.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0]));
-  });
-
-  it('3D', () => {
-    const t = K.zerosLike(tfc.randomUniform([2, 2, 2], -10, 10));
-    expect(K.shape(t.read())).toEqual([2, 2, 2]);
-    expect(t.read().dataSync()).toEqual(new Float32Array([
-      0, 0, 0, 0, 0, 0, 0, 0
-    ]));
-  });
-
-  it('4D', () => {
-    const q = K.zerosLike(tfc.randomUniform([1, 2, 1, 3], -10, 10));
-    expect(K.shape(q.read())).toEqual([1, 2, 1, 3]);
-    expect(q.read().dataSync()).toEqual(new Float32Array([0, 0, 0, 0, 0, 0]));
-  });
-});
-
-describeMathCPUAndGPU('OnesLike', () => {
-  it('Scalar', () => {
-    const s = K.onesLike(tfc.randomUniform([], -10, 10));
-    expect(K.shape(s.read())).toEqual([]);
-    expect(s.read().dataSync()).toEqual(new Float32Array([1]));
-  });
-
-  it('Vector', () => {
-    const v = K.onesLike(tfc.randomUniform([3], -10, 10));
-    expect(K.shape(v.read())).toEqual([3]);
-    expect(v.read().dataSync()).toEqual(new Float32Array([1, 1, 1]));
-  });
-
-  it('Matrix', () => {
-    const m = K.onesLike(tfc.randomUniform([2, 2], -10, 10));
-    expect(K.shape(m.read())).toEqual([2, 2]);
-    expect(m.read().dataSync()).toEqual(new Float32Array([1, 1, 1, 1]));
-  });
-
-  it('3D', () => {
-    const t = K.onesLike(tfc.randomUniform([2, 2, 2], -10, 10));
-    expect(K.shape(t.read())).toEqual([2, 2, 2]);
-    expect(t.read().dataSync()).toEqual(new Float32Array([
-      1, 1, 1, 1, 1, 1, 1, 1
-    ]));
-  });
-
-  it('4D', () => {
-    const q = K.onesLike(tfc.randomUniform([1, 2, 1, 3], -10, 10));
-    expect(K.shape(q.read())).toEqual([1, 2, 1, 3]);
-    expect(q.read().dataSync()).toEqual(new Float32Array([1, 1, 1, 1, 1, 1]));
-  });
-});
-
 describeMathCPUAndGPU('Identity', () => {
   it('Scalar', () => {
     const s = scalar(12);
@@ -1072,20 +865,6 @@ describeMathCPUAndGPU('eye (I-matrix builder)', () => {
     const I = K.eye(2);
     expect(I.shape).toEqual([2, 2]);
     expect(I.dataSync()).toEqual(new Float32Array([1, 0, 0, 1]));
-  });
-  it('Variable Zero sized 2D matrix', () => {
-    const I = K.eyeVariable(0);
-    expect(I.shape).toEqual([0, 0]);
-  });
-  it('Variable 1 sized 2D matrix', () => {
-    const I = K.eyeVariable(1);
-    expect(I.shape).toEqual([1, 1]);
-    expect(I.read().dataSync()).toEqual(new Float32Array([1]));
-  });
-  it('Variable 2 sized 2D matrix', () => {
-    const I = K.eyeVariable(2);
-    expect(I.shape).toEqual([2, 2]);
-    expect(I.read().dataSync()).toEqual(new Float32Array([1, 0, 0, 1]));
   });
 });
 
@@ -1137,40 +916,6 @@ describeMathCPUAndGPU('randomNormal', () => {
       expect(K.shape(y)).toEqual([2, 3, 4]);
     });
   }
-});
-
-describeMathCPUAndGPU('Variable update', () => {
-  it('Update', () => {
-    const v = new LayerVariable(scalar(10.0));
-    K.update(v, scalar(20.0));
-    expectTensorsClose(v.read(), scalar(20.0));
-  });
-  it('Update: Incompatible shape', () => {
-    const v = new LayerVariable(tensor1d([10.0, 20.0]));
-    const x = tensor1d([10.0, 20.0, 30.0]);
-    expect(() => K.update(v, x)).toThrowError();
-  });
-  it('UpdateAdd', () => {
-    const v = new LayerVariable(scalar(10.0));
-    K.updateAdd(v, scalar(20.0));
-    expectTensorsClose(v.read(), scalar(30.0));
-  });
-  it('UpdateAdd: Incompatible shape', () => {
-    const v = new LayerVariable(tensor1d([10.0, 20.0]));
-    const x = tensor1d([0.0, 10.0, 20.0]);
-    expect(() => K.updateAdd(v, x)).toThrowError();
-  });
-  it('UpdateSub', () => {
-    const v = new LayerVariable(scalar(10.0));
-    K.updateSub(v, scalar(20.0));
-    const vNew = v.read();
-    expectTensorsClose(vNew, scalar(-10.0));
-  });
-  it('UpdateSub: Incompatible shape', () => {
-    const v = new LayerVariable(tensor1d([10.0, 20.0]));
-    const x = tensor1d([0.0, 10.0, 20.0]);
-    expect(() => K.updateSub(v, x)).toThrowError();
-  });
 });
 
 describeMathCPUAndGPU('dot', () => {
