+++ src/variables.ts
@@ -9,9 +9,12 @@
  */
 
 import * as tfc from '@tensorflow/tfjs-core';
+import {Tensor} from '@tensorflow/tfjs-core';
 
+import {eye, randomNormal} from './backend/tfjs_backend';
 import {getScopedTensorName, getUniqueTensorName} from './common';
 import {Constraint} from './constraints';
+import {NotImplementedError} from './errors';
 import {DType, getNextUniqueTensorId, Shape, SymbolicTensor} from './types';
 
 const DEFAULT_VARIABLE_NAME_PREFIX = 'Variable';
@@ -51,7 +54,7 @@ export class LayerVariable {
    * @throws ValueError if `name` is `null` or `undefined`.
    */
   constructor(
-      val: tfc.Tensor, dtype: DType = DType.float32,
+      val: Tensor, dtype: DType = DType.float32,
       name = DEFAULT_VARIABLE_NAME_PREFIX, trainable = true,
       constraint: Constraint = null) {
     this.dtype = dtype == null ? DType.float32 : dtype;
@@ -75,7 +78,7 @@ export class LayerVariable {
    * the invocation. Future mutations in the value of the tensor will only
    * be reflected by future calls to this method.
    */
-  read(): tfc.Tensor {
+  read(): Tensor {
     return this.val;
   }
 
@@ -86,7 +89,7 @@ export class LayerVariable {
    *   dtype and shape of the Variable.
    * @return This Variable.
    */
-  write(newVal: tfc.Tensor) {
+  write(newVal: Tensor) {
     // TODO(cais): Once  TF.js Core supports Tensor.dtype, check dtype match.
     checkShapesMatch(this.val, newVal);
     this.val.assign(newVal);
@@ -98,10 +101,204 @@ export class LayerVariable {
 }
 
 function checkShapesMatch(
-    x: tfc.Tensor|SymbolicTensor, y: tfc.Tensor|SymbolicTensor): void {
+    x: Tensor|SymbolicTensor, y: Tensor|SymbolicTensor): void {
   if (x.shape.toString() !== y.shape.toString()) {
     throw new Error(
         'Shape mismatch: ' + JSON.stringify(x.shape) + ' vs. ' +
         JSON.stringify(y.shape));
   }
+}
+
+/**
+ * Create a Variable.
+ * @param x The initial value of the `Variable`.
+ * @param dtype optional, the type of the variable.
+ * @param name optional, the name of the variable, default provided by
+ * Variable.
+ * @param constraint optional, a constraint to be applied after every update.
+ * @return The newly instantiated `Variable`.
+ */
+export function variable(
+    x: Tensor, dtype?: DType, name?: string,
+    constraint?: Constraint): LayerVariable {
+  return new LayerVariable(x, dtype, name, true, constraint);
+}
+
+/**
+ * Instantiates an all-zeros Variable and returns it.
+ *
+ * @param shape Shape of the tensor.
+ * @param dtype DType of the tensor.
+ * @param name Name of the tensor.
+ * @return An all-zero Variable.
+ */
+export function zerosVariable(
+    shape: Shape, dtype?: DType, name?: string): LayerVariable {
+  // TODO(cais): Implement logic for dtype.
+  return new LayerVariable(tfc.zeros(shape), dtype, name);
+}
+
+/**
+ * Instantiates an all-zeros tensor of the same shape as another tensor.
+ *
+ * @param x The other tensor.
+ * @param dtype DType of the tensor.
+ * @param name Name of the tensor.
+ * @return A newly instantiated Variable.
+ */
+export function zerosLike(
+    x: Tensor, dtype?: DType, name?: string): LayerVariable {
+  return new LayerVariable(tfc.zerosLike(x), dtype, name);
+}
+
+/**
+ * Instantiates an all-ones tensor and returns it.
+ *
+ * @param shape Shape of the tensor.
+ * @param dtype DType of the tensor.
+ * @param name Name of the tensor.
+ * @return An all-ones Variable.
+ */
+export function onesVariable(
+    shape: Shape, dtype?: DType, name?: string): LayerVariable {
+  // TODO(cais): Implement logic for dtype.
+  const allocated = tfc.ones(shape);
+  return new LayerVariable(allocated, dtype, name);
+}
+
+/**
+ * Instantiates an all-ones tensor of the same shape as another tensor.
+ *
+ * @param x The other tensor.
+ * @param dtype DType of the tensor.
+ * @param name Name of the tensor.
+ * @return A newly instantiated Variable.
+ */
+export function onesLike(
+    x: Tensor, dtype?: DType, name?: string): LayerVariable {
+  const allocated = tfc.onesLike(x);
+  return new LayerVariable(allocated, dtype, name);
+}
+
+/**
+ * Instantiate an identity matrix and returns it, as a Variable
+ *
+ * @param size Number of rows/columns.
+ * @param dtype Data type of returned Variable.
+ * @param name Name of returned Variable.
+ * @return A Variable, an identity matrix.
+ */
+export function eyeVariable(
+    size: number, dtype?: DType, name?: string): LayerVariable {
+  return new LayerVariable(eye(size, dtype), dtype, name);
+}
+/**
+ * Get a Variable with uniform distribution of values.
+ * @param shape Shape of the tensor.
+ * @param minval Lower bound of the uniform distribution.
+ * @param maxval Upper bound of the uniform distribution.
+ * @param dtype
+ * @param seed
+ * @param name Optional name.
+ * @return The uniform-random Variable.
+ */
+export function randomUniformVariable(
+    shape: Shape, minval: number, maxval: number, dtype?: DType, seed?: number,
+    name = 'randomUniform'): LayerVariable {
+  return new LayerVariable(
+      tfc.randomUniform(shape, minval, maxval, dtype), dtype, name);
+}
+
+/**
+ * Get a Variable with truncated-normal distribution of values.
+ * @param shape Shape of the tensor.
+ * @param mean mean value of the normal distribution.
+ * @param stddev standard deviation of the normal distribution.
+ * @param dtype
+ * @param seed
+ * @param name Optional name.
+ * @return The truncated-normal-random Variable.
+ */
+export function truncatedNormalVariable(
+    shape: Shape, mean = 0.0, stddev = 1.0, dtype?: DType, seed?: number,
+    name = 'truncatedNormal'): LayerVariable {
+  // TODO(cais): Implement logic for dtype and seed once they are supported
+  // by deeplearn.js.
+  if (dtype === DType.bool) {
+    throw new NotImplementedError(`randomNormal does not support dType bool.`);
+  }
+  return new LayerVariable(
+      tfc.truncatedNormal(shape, mean, stddev, dtype, seed), dtype, name);
+}
+/**
+ * Get a Variable with normal distribution of values.
+ * @param shape Shape of the tensor.
+ * @param mean mean value of the normal distribution.
+ * @param stddev standard deviation of the normal distribution.
+ * @param dtype
+ * @param seed
+ * @param name Optional name.
+ * @return The truncated-normal-random Variable.
+ */
+export function randomNormalVariable(
+    shape: Shape, mean = 0.0, stddev = 1.0, dtype?: DType, seed?: number,
+    name = 'randomNormal'): LayerVariable {
+  return new LayerVariable(
+      randomNormal(shape, mean, stddev, dtype, seed), dtype, name);
+}
+
+/**
+ * Update the value of a Variable.
+ * @param x The Variable to be updated.
+ * @param xNew The new value to update to.
+ * @return The Variable updated.
+ */
+export function update(x: LayerVariable, xNew: Tensor): LayerVariable {
+  return x.write(xNew);
+}
+
+/**
+ * Update the value of a Variable by adding an increment.
+ * @param x The Variable to be updated.
+ * @param increment The incrment to add to `x`.
+ * @return The Variable updated.
+ */
+export function updateAdd(x: LayerVariable, increment: Tensor): LayerVariable {
+  return x.write(tfc.add(x.read(), increment));
+}
+
+/**
+ * Update the value of a Variable by subtracting a decrement.
+ * @param x The Variable to be updated.
+ * @param decrement The decrement to subtract from `x`.
+ * @return The Variable updated.
+ */
+export function updateSub(x: LayerVariable, decrement: Tensor): LayerVariable {
+  return x.write(tfc.sub(x.read(), decrement));
+}
+
+/**
+ * Get the values of an array of Variables.
+ *
+ * @param tensors An `Array` of `Variable`s to get the values of.
+ * @return The values of the inputs, as an `Array` of `Tensor`s.
+ */
+export function batchGetValue(xs: LayerVariable[]): Tensor[] {
+  return xs.map(x => x.read());
+}
+
+/**
+ * Update the value of multiple Variables at once.
+ *
+ * @param variablesAndValues An `Array`, each element is of type
+ *   [Variable, Tensor]. The first item is the
+ *   `Variable` of which the value is to be updated. The second item
+ *   carries the new value.
+ */
+export function batchSetValue(
+    variablesAndValues: Array<[LayerVariable, Tensor]>): void {
+  variablesAndValues.map((variableAndValue) => {
+    const variable: LayerVariable = variableAndValue[0];
+    variable.write(variableAndValue[1]);
+  });
 }
